+++ cryptofeed/connection_handler.py
@@ -23,16 +23,18 @@ LOG = logging.getLogger('feedhandler')
 
 
 class ConnectionHandler:
-    def __init__(self, conn: AsyncConnection, subscribe: Awaitable, handler: Awaitable, retries: int, timeout=120, timeout_interval=30, exceptions=None, log_on_error=False):
+    def __init__(self, conn: AsyncConnection, subscribe: Awaitable, handler: Awaitable, authenticate: Awaitable, retries: int, timeout=120, timeout_interval=30, exceptions=None, log_on_error=False, start_delay=0):
         self.conn = conn
         self.subscribe = subscribe
         self.handler = handler
+        self.authenticate = authenticate
         self.retries = retries
         self.exceptions = exceptions
         self.log_on_error = log_on_error
         self.timeout = timeout
         self.timeout_interval = timeout_interval
         self.running = True
+        self.start_delay = start_delay
 
     def start(self, loop: asyncio.AbstractEventLoop):
         loop.create_task(self._create_connection())
@@ -47,6 +49,7 @@ class ConnectionHandler:
             await asyncio.sleep(self.timeout_interval)
 
     async def _create_connection(self):
+        await asyncio.sleep(self.start_delay)
         retries = 0
         rate_limited = 1
         delay = 1
@@ -57,6 +60,7 @@ class ConnectionHandler:
                     retries = 0
                     rate_limited = 0
                     delay = 1
+                    await self.authenticate(connection)
                     await self.subscribe(connection)
                     if self.timeout != -1:
                         loop = asyncio.get_running_loop()